requirements.txt
elasticsearch
requests
newsapi-python
